backend/requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.4.60
imgbeddings==0.0.4
numpy==1.21.3
Pillow==8.4.0
pymongo==3.12.0
Flask-Cors==3.0.10
